+++ src/main/java/com/example/demo/dao/CourseDAO.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.springframework.stereotype.Repository;
 
 import com.example.demo.model.Course;
+import com.example.demo.model.CourseSubscribedVideo;
 
 @Repository
 public interface CourseDAO  {
