+++ src/main/java/com/example/demo/dao/courseDAOImp.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import com.example.demo.model.Course;
+import com.example.demo.model.CourseSubscribedVideo;
 @Repository
 public class courseDAOImp implements CourseDAO {
